+++ apps/client/src/app/components/holding-detail-dialog/holding-detail-dialog.component.ts
@@ -63,7 +63,7 @@ import { PageEvent } from '@angular/material/paginator';
 import { SortDirection } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatTabsModule } from '@angular/material/tabs';
-import { Router, RouterModule } from '@angular/router';
+import { NavigationStart, Router, RouterModule } from '@angular/router';
 import { IonIcon } from '@ionic/angular/standalone';
 import { Account, MarketData, Tag } from '@prisma/client';
 import { isUUID } from 'class-validator';
@@ -80,9 +80,12 @@ import {
 } from 'ionicons/icons';
 import { isNumber, round, uniqBy } from 'lodash';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
-import { switchMap } from 'rxjs/operators';
+import { filter, switchMap } from 'rxjs/operators';
 
-import { HoldingDetailDialogParams } from './interfaces/interfaces';
+import {
+  HoldingDetailDialogParams,
+  HoldingDetailDialogResult
+} from './interfaces/interfaces';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -186,9 +189,10 @@ export class GfHoldingDetailDialogComponent implements OnInit {
   protected value: number;
 
   protected readonly data = inject<HoldingDetailDialogParams>(MAT_DIALOG_DATA);
-  protected readonly dialogRef = inject(
-    MatDialogRef<GfHoldingDetailDialogComponent>
-  );
+  protected readonly dialogRef =
+    inject<
+      MatDialogRef<GfHoldingDetailDialogComponent, HoldingDetailDialogResult>
+    >(MatDialogRef);
 
   private tags: Tag[];
 
@@ -200,6 +204,17 @@ export class GfHoldingDetailDialogComponent implements OnInit {
   private readonly userService = inject(UserService);
 
   public constructor() {
+    this.router.events
+      .pipe(
+        filter((event) => {
+          return event instanceof NavigationStart;
+        }),
+        takeUntilDestroyed(this.destroyRef)
+      )
+      .subscribe(() => {
+        this.dialogRef.close({ isNavigating: true });
+      });
+
     addIcons({
       arrowDownCircleOutline,
       createOutline,
@@ -589,17 +604,6 @@ export class GfHoldingDetailDialogComponent implements OnInit {
     this.fetchActivities();
   }
 
-  protected onCloneActivity(aActivity: Activity) {
-    this.router.navigate(
-      internalRoutes.portfolio.subRoutes.activities.routerLink,
-      {
-        queryParams: { activityId: aActivity.id, createDialog: true }
-      }
-    );
-
-    this.dialogRef.close();
-  }
-
   protected onClose() {
     this.dialogRef.close();
   }
@@ -659,17 +663,6 @@ export class GfHoldingDetailDialogComponent implements OnInit {
     if (withRefresh) {
       this.fetchMarketData();
     }
-  }
-
-  protected onUpdateActivity(aActivity: Activity) {
-    this.router.navigate(
-      internalRoutes.portfolio.subRoutes.activities.routerLink,
-      {
-        queryParams: { activityId: aActivity.id, editDialog: true }
-      }
-    );
-
-    this.dialogRef.close();
   }
 
   private fetchActivities(filters: Filter[] = this.getActivityFilters()) {